# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trie backend — Hybrid RAG",
    "description": "멀티모달 + 에이전트 기반 하이브리드 RAG 통합 검색 시스템 (BE)",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/kakao": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Kakao Login",
        "description": "카카오 인가 코드로 로그인 → 사용자 동기화 → JWT 발급.",
        "operationId": "kakao_login_auth_kakao_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KakaoLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Refresh 토큰으로 Access 토큰 재발급 (Silent Refresh).",
        "operationId": "refresh_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "현재 사용자의 리프레시 토큰을 폐기.",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Logout Auth Logout Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/documents": {
      "post": {
        "tags": [
          "documents(통신)"
        ],
        "summary": "Upload Documents",
        "description": "비정형 문서(PDF/DOCX) 멀티 업로드 → 검증·저장·수집 + 업로더 메타 동기화(F7).",
        "operationId": "upload_documents_documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documents(통신)"
        ],
        "summary": "List Documents",
        "description": "문서 목록(최신순, domain 필터 선택).",
        "operationId": "list_documents_documents_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "documents(통신)"
        ],
        "summary": "Get Document",
        "description": "문서 단건 조회.",
        "operationId": "get_document_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents(통신)"
        ],
        "summary": "Delete Document",
        "description": "문서 + ChromaDB 벡터 + 저장 파일 동기 삭제.",
        "operationId": "delete_document_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Document Documents  Document Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/ingest": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Ingest Document",
        "operationId": "ingest_document_api_v1_documents_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_document_api_v1_documents_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "AccessTokenResponse"
      },
...
      "Body_ingest_document_api_v1_documents_ingest_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "etc"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingest_document_api_v1_documents_ingest_post"
      },
...
      "Body_upload_documents_documents_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": "etc"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_documents_post"
      },
...
      "ChunkPreview": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "token_estimate": {
            "type": "integer",
            "title": "Token Estimate"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "index",
          "char_count",
          "token_estimate",
          "text"
        ],
        "title": "ChunkPreview"
      },
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "type": "string",
            "title": "Doc Type"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "doc_type",
          "domain",
          "status",
          "char_count",
          "chunk_count",
          "created_at"
        ],
        "title": "DocumentOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResponse": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "type": "string",
            "title": "Doc Type"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "elapsed_sec": {
            "type": "number",
            "title": "Elapsed Sec"
          },
          "chunk_preview": {
            "items": {
              "$ref": "#/components/schemas/ChunkPreview"
            },
            "type": "array",
            "title": "Chunk Preview"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "doc_type",
          "domain",
          "status",
          "char_count",
          "chunk_count",
          "elapsed_sec",
          "chunk_preview"
        ],
        "title": "IngestResponse"
      },
      "KakaoLoginRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "redirect_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect Uri"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "KakaoLoginRequest"
      },
      "MultiUploadResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UploadedDocument"
            },
            "type": "array",
            "title": "Items"
          },
          "failed": {
            "items": {
              "$ref": "#/components/schemas/UploadFailure"
            },
            "type": "array",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "items",
          "failed"
        ],
        "title": "MultiUploadResponse"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse"
      },
      "UploadFailure": {
        "properties": {
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "UploadFailure"
      },
      "UploadedDocument": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "type": "string",
            "title": "Doc Type"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "doc_type",
          "domain",
          "status",
          "chunk_count"
        ],
        "title": "UploadedDocument"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "provider"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}